move SDL3_image out SDL3input
