+++ Core/GameEngineDevice/CMakeLists.txt
@@ -197,8 +197,10 @@ if(SAGE_USE_SDL3 AND NOT IS_VS6_BUILD)
     list(APPEND GAMEENGINEDEVICE_SRC
         Include/SDL3GameEngine.h
         Include/SDL3Device/GameClient/SDL3Input.h
+        Include/SDL3Device/GameClient/SDL3Cursor.h
         Source/SDL3GameEngine.cpp
         Source/SDL3Device/GameClient/SDL3Input.cpp
+        Source/SDL3Device/GameClient/SDL3Cursor.cpp
     )
 endif()
